+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # PyMagento Changelog
 
+## Unreleased
+
+* Minor type hints improvements
+
 ## 1.0.0 (2021/12/21)
 
 First public release.